+++ tests/EntityServiceTestAsset/MyEntity.php
@@ -11,7 +11,15 @@ namespace PolderKnowledge\EntityServiceTestAsset;
 
 class MyEntity
 {
+    private $id;
+
+    public function __construct($id = null)
+    {
+        $this->id = $id;
+    }
+
     public function getId()
     {
+        return $this->id;
     }
 }